moving shap over from an eval file to inference file
